+++ web/documentserver-example/csharp-mvc/WebEditor.ashx.cs
@@ -243,7 +243,13 @@ namespace OnlineEditorsExampleMVC
                 var fileUri = DocManagerHelper.GetDownloadUrl(fileName);
 
                 var extension = (Path.GetExtension(fileName).ToLower() ?? "").Trim('.');
-                var internalExtension = "ooxml";
+                string conversionExtension = "ooxml";
+                object fileExt;
+
+                if (body.TryGetValue("fileExt", out fileExt) && !String.IsNullOrEmpty(fileExt.ToString()))
+                {
+                    conversionExtension = fileExt.ToString();
+                }
 
                 // check if the file with such an extension can be converted
                 if (DocManagerHelper.ConvertExts.Contains("." + extension))
@@ -261,7 +267,7 @@ namespace OnlineEditorsExampleMVC
 
                     // get the url and file type of the converted file
                     Dictionary<string, string> newFileData;
-                    var result = ServiceConverter.GetConvertedData(downloadUri.ToString(), extension, internalExtension, key, true, out newFileData, filePass, lang);
+                    var result = ServiceConverter.GetConvertedData(downloadUri.ToString(), extension, conversionExtension, key, true, out newFileData, filePass, lang);
                     if (result != 100)
                     {
                         context.Response.Write("{ \"step\" : \"" + result + "\", \"filename\" : \"" + fileName + "\"}");
@@ -396,8 +402,17 @@ namespace OnlineEditorsExampleMVC
             context.Response.ContentType = "text/plain";
             try
             {
-                var fileName = Path.GetFileName(context.Request["fileName"]);
-                Remove(fileName);  // remove a file and its history if it exists
+                string fileName = context.Request["fileName"];
+
+                if (!String.IsNullOrEmpty(fileName))
+                {
+                    fileName = Path.GetFileName(context.Request["fileName"]);
+                    Remove(fileName);  // remove a file and its history if it exists
+                }
+                else
+                {
+                    RemoveUserDirectory(); // remove the user's directory
+                }
 
                 context.Response.Write("{ \"success\": true }");
             }
@@ -415,6 +430,14 @@ namespace OnlineEditorsExampleMVC
 
             if (File.Exists(path)) File.Delete(path);
             if (Directory.Exists(histDir)) Directory.Delete(histDir, true);
+        }
+
+        // remove the user's directory
+        private static void RemoveUserDirectory()
+        {
+            var path = DocManagerHelper.StoragePath("", null);  // get the path to the user directory
+
+            if (Directory.Exists(path)) Directory.Delete(path, true);
         }
 
         // get files information