+++ web/documentserver-example/java/src/main/webapp/scripts/jscript.js
@@ -108,7 +108,7 @@ if (typeof jQuery !== "undefined") {
     });
     
     var timer = null;
-    var checkConvert = function (filePass) {
+    var checkConvert = function (filePass, fileType) {
 		filePass = filePass ? filePass : null;
         if (timer !== null) {
             clearTimeout(timer);
@@ -136,7 +136,7 @@ if (typeof jQuery !== "undefined") {
                 contentType: "text/xml",
                 type: "post",
                 dataType: "json",
-                data: JSON.stringify({filename: fileName, filePass: filePass}),
+                data: JSON.stringify({filename: fileName, filePass: filePass, fileExt: fileType}),
                 url: UrlConverter,
                 complete: function (data) {
                     var responseText = data.responseText;
@@ -152,6 +152,12 @@ if (typeof jQuery !== "undefined") {
                             }
                             return;
                         } else {
+                            if (response.error.includes("Error conversion output format")){
+                                jq("#select-file-type").removeClass("invisible");
+                                jq("#step2").removeClass("current");
+                                jq("#hiddenFileName").attr("placeholder",filePass);
+                                return;
+                            }
                             jq(".current").removeClass("current");
                             jq(".step:not(.done)").addClass("error");
                             jq("#mainProgress .error-message").show().find("span").text(response.error);
@@ -163,7 +169,7 @@ if (typeof jQuery !== "undefined") {
                     jq("#hiddenFileName").val(response.filename);
 
                     if (response.step && response.step < 100) {
-                        checkConvert(filePass);
+                        checkConvert(filePass, fileType);
                     } else {
                         jq("#step2").addClass("done").removeClass("current");
                         loadScripts();
@@ -232,6 +238,15 @@ if (typeof jQuery !== "undefined") {
             setCookie("ulang", langSel.val());
         });
     };
+
+    jq(document).on("click", ".file-type:not(.disable)", function () {
+        const currentElement = jq(this);
+        var fileType = currentElement.attr("data");
+        var filePass = jq("#hiddenFileName").attr("placeholder");
+        jq('.file-type').addClass(["disable", "pale"]);
+        currentElement.removeClass("pale");
+        checkConvert(filePass, fileType);
+    });
 
     jq(document).on("click", "#enterPass", function () {
         var pass = jq("#filePass").val();
@@ -309,6 +324,21 @@ if (typeof jQuery !== "undefined") {
                 document.location.reload();
             }
         });
+    });
+
+    jq(document).on("click", ".clear-all", function () {
+        if (confirm("Delete all the files?")) {
+            jq.ajax({
+                async: true,
+                contentType: "text/xml",
+                url: "IndexServlet?type=remove",
+                complete: function (data) {
+                    if (JSON.parse(data.responseText).success) {
+                        window.location.reload(true);
+                    }
+                }
+            });
+        }
     });
 
         function showUserTooltip (isMobile) {