+++ web/documentserver-example/python/src/utils/docManager.py
@@ -286,6 +286,16 @@ def removeFile(filename, req):
         shutil.rmtree(histDir)
 
 
+# remove the user's directory and all the containing files
+def removeUserFolder(req):
+    path = os.path.normpath(getRootFolder(req))
+
+    if not path.startswith(str(config_manager.storage_path())) or not os.path.exists(path):
+        raise FileNotFoundError
+
+    shutil.rmtree(path)
+
+
 # generate file key
 def generateFileKey(filename, req):
     path = getStoragePath(filename, req)